DRY up configs
