+++ signature-aggregator/config/keys.go
@@ -16,6 +16,4 @@ const (
 	APIPortKey            = "api-port"
 	MetricsPortKey        = "metrics-port"
 	SignatureCacheSizeKey = "signature-cache-size"
-	TLSCertPathKey        = "tls-cert-path"
-	TLSKeyPathKey         = "tls-key-path"
 )
