+++ signature-aggregator/config/viper.go
@@ -4,15 +4,12 @@
 package config
 
 import (
-	"crypto/tls"
 	"fmt"
-	"os"
 	"strings"
 
-	"github.com/ava-labs/avalanchego/staking"
+	commonConfig "github.com/ava-labs/icm-services/config"
 	"github.com/spf13/pflag"
 	"github.com/spf13/viper"
-	"go.uber.org/zap"
 )
 
 func NewConfig(v *viper.Viper) (Config, error) {
@@ -80,9 +77,8 @@ func BuildConfig(v *viper.Viper) (Config, error) {
 		return cfg, fmt.Errorf("failed to unmarshal viper config: %w", err)
 	}
 
-	//
-	if v.IsSet(TLSKeyPathKey) || v.IsSet(TLSCertPathKey) {
-		cert, err := getTLSCertFromFile(v)
+	if v.IsSet(commonConfig.TLSKeyPathKey) || v.IsSet(commonConfig.TLSCertPathKey) {
+		cert, err := commonConfig.GetTLSCertFromFile(v)
 		if err != nil {
 			return cfg, fmt.Errorf("failed to initialize TLS certificate: %w", err)
 		}
@@ -90,50 +86,4 @@ func BuildConfig(v *viper.Viper) (Config, error) {
 	}
 
 	return cfg, nil
-}
-
-func getTLSCertFromFile(v *viper.Viper) (tls.Certificate, error) {
-	if !v.IsSet(TLSKeyPathKey) || !v.IsSet(TLSCertPathKey) {
-		return tls.Certificate{}, fmt.Errorf("TLS key or cert path not set")
-	}
-	// Parse the staking key/cert paths and expand environment variables
-	keyPath := getExpandedPath(v, TLSKeyPathKey)
-	certPath := getExpandedPath(v, TLSCertPathKey)
-
-	var keyMissing, certMissing bool
-
-	if _, err := os.Stat(keyPath); os.IsNotExist(err) {
-		keyMissing = true
-	}
-	if _, err := os.Stat(certPath); os.IsNotExist(err) {
-		certMissing = true
-	}
-	if !(keyMissing && certMissing) && (keyMissing || certMissing) {
-		// If only one of the key/cert pair is missing return an error
-		// otherwise, create the staking key/cert pair
-		return tls.Certificate{}, fmt.Errorf("TLS key or cert file is missing from configured path.", zap.String("keyPath", keyPath), zap.String("certPath", certPath))
-	} else if keyMissing && certMissing {
-		// Create the key/cert pair if [TLSKeyPath] and [TLSCertPath] are set but the files are missing
-		if err := staking.InitNodeStakingKeyPair(keyPath, certPath); err != nil {
-			return tls.Certificate{}, fmt.Errorf("couldn't generate TLS key/cert: %w", err)
-		}
-	}
-
-	// Load and parse the staking key/cert
-	cert, err := staking.LoadTLSCertFromFiles(keyPath, certPath)
-	if err != nil {
-		return tls.Certificate{}, fmt.Errorf("couldn't read staking certificate: %w", err)
-	}
-	return *cert, nil
-}
-
-// getExpandedPath gets the string in viper corresponding to [key] and expands
-// any variables using the OS env.
-func getExpandedPath(v *viper.Viper, key string) string {
-	return os.Expand(
-		v.GetString(key),
-		func(strVar string) string {
-			return os.Getenv(strVar)
-		},
-	)
 }
